Update SpoofClientPatch.kt to fix reddit is fun
https://old.reddit.com/r/redditdev/comments/1d3spoc/error_getting_submitted_with_mobile_useragent/

https://github.com/Docile-Alligator/Infinity-For-Reddit/commit/07c9865ab2b60135604a84f0bbb9d460d7eabee7

might be temp: https://github.com/ReVanced/revanced-patches/pull/3253#pullrequestreview-2089382033

"oSumAtrIX left a comment
This PR was made because Reddit broke their oAuth API for the Android platform. In the past Reddit has caused the same issue and every time it resolved itself after a while. Sometimes a desktop browser user agent was required to bypass issues with logging in for example. For that reason this PR might be obsolete soon, if Reddit fixes their issue. Until then, Reddit should be contacted first and asked about the issue."